+++ src/output/Form.ts
@@ -1,3 +1,4 @@
+import type Project from '@models/Project';
 import { getFirstName } from '../locale/Locale';
 import type Locales from '../locale/Locales';
 import StructureValue from '../values/StructureValue';
@@ -5,6 +6,9 @@ import type Value from '../values/Value';
 import { toNumber } from './Stage';
 import Valued, { getOutputInputs } from './Valued';
 import { PX_PER_METER } from './outputToCSS';
+import toStructure from '@basis/toStructure';
+import { getBind } from '@locale/getBind';
+import { TYPE_SYMBOL } from '@parser/Symbols';
 
 /** This is a wrapper class for a Form value, which represents some kind of shape that's used as a collision boundary. */
 export abstract class Form extends Valued {
@@ -36,6 +40,24 @@ export abstract class Form extends Valued {
     abstract getDescription(locales: Locales): string;
 }
 
+export function createFormType(locales: Locales) {
+    return toStructure(`
+    ${getBind(locales, (locale) => locale.output.Form, TYPE_SYMBOL)}()
+`);
+}
+
+export function createRectangleType(locales: Locales) {
+    return toStructure(`
+    ${getBind(locales, (locale) => locale.output.Rectangle, TYPE_SYMBOL)} Form (
+        ${getBind(locales, (locale) => locale.output.Rectangle.left)}•#m
+        ${getBind(locales, (locale) => locale.output.Rectangle.top)}•#m
+        ${getBind(locales, (locale) => locale.output.Rectangle.right)}•#m
+        ${getBind(locales, (locale) => locale.output.Rectangle.bottom)}•#m
+        ${getBind(locales, (locale) => locale.output.Rectangle.z)}•#m: 0m
+    )
+`);
+}
+
 export class Rectangle extends Form {
     readonly left: number;
     readonly top: number;
@@ -49,7 +71,7 @@ export class Rectangle extends Form {
         top: number,
         right: number,
         bottom: number,
-        z: number
+        z: number,
     ) {
         super(value);
 
@@ -107,9 +129,89 @@ export class Rectangle extends Form {
     }
 }
 
-export function toRectangle(value: Value | undefined) {
+export function createCircleType(locales: Locales) {
+    return toStructure(`
+    ${getBind(locales, (locale) => locale.output.Circle, TYPE_SYMBOL)} Form (
+        ${getBind(locales, (locale) => locale.output.Circle.radius)}•#m
+        ${getBind(locales, (locale) => locale.output.Circle.x)}•#m
+        ${getBind(locales, (locale) => locale.output.Circle.y)}•#m
+        ${getBind(locales, (locale) => locale.output.Circle.z)}•#m: 0m
+    )
+`);
+}
+
+export class Circle extends Form {
+    readonly x: number;
+    readonly y: number;
+    readonly z: number;
+    readonly radius: number;
+
+    constructor(value: Value, x: number, y: number, z: number, radius: number) {
+        super(value);
+
+        this.x = x;
+        this.y = y;
+        this.z = z;
+        this.radius = radius;
+    }
+
+    getLeft() {
+        return this.x - this.radius;
+    }
+
+    getTop() {
+        return this.y + this.radius;
+    }
+
+    getZ() {
+        return this.z;
+    }
+
+    getWidth() {
+        return this.radius * 2;
+    }
+
+    getHeight() {
+        return this.radius * 2;
+    }
+
+    getCoordinates() {
+        return {
+            x: this.x * PX_PER_METER,
+            y: -this.y * PX_PER_METER,
+            radius: this.radius * PX_PER_METER,
+        };
+    }
+
+    toCSSClip() {
+        const { x, y, radius } = this.getCoordinates();
+        return `circle(${radius}px ${y}px ${x}px)`;
+    }
+
+    toSVGPath() {
+        const { x, y, radius } = this.getCoordinates();
+        return `M ${x - radius}, ${y} a ${radius},${radius} 0 1,0 ${
+            radius * 2
+        },0 a ${radius},${radius} 0 1,0 ${-radius * 2},0`;
+    }
+
+    getDescription(locales: Locales): string {
+        return locales.get((l) => getFirstName(l.output.Rectangle.names));
+    }
+}
+
+export function toForm(
+    project: Project,
+    value: Value | undefined,
+): Form | undefined {
     if (!(value instanceof StructureValue)) return undefined;
 
+    if (value.is(project.shares.output.Rectangle)) return toRectangle(value);
+    else if (value.is(project.shares.output.Circle)) return toCircle(value);
+    else return undefined;
+}
+
+export function toRectangle(value: StructureValue) {
     const [leftVal, topVal, rightVal, bottomVal, zVal] = getOutputInputs(value);
 
     const left = toNumber(leftVal);
@@ -122,5 +224,19 @@ export function toRectangle(value: Value | undefined) {
         right !== undefined &&
         bottom !== undefined
         ? new Rectangle(value, left, top, right, bottom, z)
+        : undefined;
+}
+
+export function toCircle(value: Value | undefined) {
+    if (!(value instanceof StructureValue)) return undefined;
+
+    const [radiusVal, xVal, yVal, zVal] = getOutputInputs(value);
+
+    const radius = toNumber(radiusVal);
+    const x = toNumber(xVal);
+    const y = toNumber(yVal);
+    const z = toNumber(zVal) ?? 0;
+    return x !== undefined && y !== undefined && radius !== undefined
+        ? new Circle(value, x, y, z, radius)
         : undefined;
 }