+++ src/output/Shape.ts
@@ -8,7 +8,7 @@ import type TextLang from './TextLang';
 import type { DefinitePose } from './Pose';
 import type Pose from './Pose';
 import type Sequence from './Sequence';
-import { Form, toRectangle } from './Form';
+import { Form, toForm } from './Form';
 import type Project from '../models/Project';
 import type Value from '../values/Value';
 import type { NameGenerator } from './Stage';
@@ -20,7 +20,10 @@ import type Locales from '../locale/Locales';
 export function createShapeType(locales: Locales) {
     return toStructure(`
     ${getBind(locales, (locale) => locale.output.Shape, TYPE_SYMBOL)} Output(
-        ${getBind(locales, (locale) => locale.output.Shape.form)}•Rectangle
+        ${getBind(
+            locales,
+            (locale) => locale.output.Shape.form,
+        )}•Rectangle|Circle
         ${getBind(locales, (locale) => locale.output.Shape.name)}•""|ø: ø
         ${getBind(locales, (locale) => locale.output.Shape.selectable)}•?: ⊥
         ${getBind(locales, (locale) => locale.output.Shape.color)}•🌈${'|ø: ø'}
@@ -49,18 +52,6 @@ export function createShapeType(locales: Locales) {
             )
             .flat()
             .join('|')}: "${DefaultStyle}"
-    )
-`);
-}
-
-export function createRectangleType(locales: Locales) {
-    return toStructure(`
-    ${getBind(locales, (locale) => locale.output.Rectangle, TYPE_SYMBOL)}(
-        ${getBind(locales, (locale) => locale.output.Rectangle.left)}•#m
-        ${getBind(locales, (locale) => locale.output.Rectangle.top)}•#m
-        ${getBind(locales, (locale) => locale.output.Rectangle.right)}•#m
-        ${getBind(locales, (locale) => locale.output.Rectangle.bottom)}•#m
-        ${getBind(locales, (locale) => locale.output.Rectangle.z)}•#m: 0m
     )
 `);
 }
@@ -168,7 +159,7 @@ export function toShape(
 ): Shape | undefined {
     if (!(value instanceof StructureValue)) return undefined;
 
-    const form = toRectangle(getOutputInput(value, 0));
+    const form = toForm(project, getOutputInput(value, 0));
 
     const {
         name,