+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 -   Improved layout of example code output refresh.
 -   Tidier spacing of project preview lists.
 
+### Added
+
+-   Added a `Circle` form, which can be added to a stage as a background.
+
 ## 0.9.394 2024-05-11
 
 ### Fixed